Test11 is remade
Test11 is remade with new name test14

+++ NetCore OOP Samples/OOP_Test14_AgainTest11/Entity/Concrete/Product.cs
@@ -12,7 +12,7 @@ namespace Entity.Concrete
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public int CategoryId { get; set; }
-        public int UnitsInStock { get; set; }
+        public short UnitsInStock { get; set; }
         public decimal UnitPrice { get; set; }
     }
 }
